refactor : items 분리
